+++ src/css/_App.scss
@@ -6,7 +6,3 @@
   margin-top: 30px;
   margin-bottom: 40px;
 }
-
-.container {
-  // margin-top: 100px;
-}
