+++ src/css/_mixins.scss
@@ -11,8 +11,10 @@
   max-width: 35px;
   display: inline-block;
   justify-content: space-around;
+  text-align: center;
   font-family: "Urbanist", sans-serif;
   font-size: 1rem;
   font-weight: bold;
+  text-decoration: none;
   cursor: pointer;
 }
