+++ src/css/index.css
@@ -60,12 +60,36 @@ body {
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  background: url(https://images3.alphacoders.com/812/812062.png) no-repeat center center fixed;
+  background: url(../img/background.jpg) no-repeat center center fixed;
   background-size: cover;
 }
+.banner .navbarContainer {
+  display: flex;
+  justify-content: space-around;
+  text-align: center;
+}
+.banner .navbarContainer .navbarButtons {
+  outline: none;
+  border: none;
+  border-radius: 15px;
+  background-color: #abdc6a;
+  color: #02afc6;
+  padding: 8px;
+  margin-top: 0;
+  margin-right: 10px;
+  width: 100px;
+  font-family: "Urbanist", sans-serif;
+  font-size: 1.6rem;
+  font-weight: bold;
+  text-decoration: none;
+  cursor: pointer;
+}
+.banner .navbarContainer .navbarButtons:hover {
+  box-shadow: 1px 1px 20px #02afc6;
+}
 .banner .header {
-  font-size: 4rem;
-  color: white;
+  font-size: 3rem;
+  color: #02afc6;
 }
 .banner .buttonBanner {
   border: none;
@@ -82,7 +106,10 @@ body {
   text-decoration: none;
   cursor: pointer;
 }
-.banner .apiLink {
+.banner .buttonBanner:hover {
+  box-shadow: 1px 1px 20px #02afc6;
+}
+.banner .buttonBanner .apiLink {
   outline: none;
   background-color: #abdc6a;
   color: #02afc6;
@@ -112,10 +139,16 @@ body {
   font-family: "Urbanist", sans-serif;
   font-size: 1rem;
   width: 300px;
+  outline: none;
+}
+.searchboxContainer .searchInput:focus {
+  outline: none;
+  border: 1px solod #abdc6a;
+  box-shadow: 1px 1px 20px #abdc6a;
 }
 
 .userCard {
-  border: 1px solid #afafaf;
+  border: 1px solid #979797;
   border-radius: 12px;
   height: 200px;
   width: 600px;
@@ -123,6 +156,9 @@ body {
   overflow: hidden;
   display: flex;
   text-align: center;
+}
+.userCard:hover {
+  box-shadow: 1px 1px 20px #02afc6;
 }
 .userCard .userCardTextContainer {
   width: 50%;
@@ -188,8 +224,10 @@ body {
   max-width: 35px;
   display: inline-block;
   justify-content: space-around;
+  text-align: center;
   font-family: "Urbanist", sans-serif;
   font-size: 1rem;
   font-weight: bold;
+  text-decoration: none;
   cursor: pointer;
 }/*# sourceMappingURL=index.css.map */
