+++ src/pages/About.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+import "../css/index.css";
+import Banner from "../components/Banner";
+
+const About = () => {
+  return (
+    <>
+      <Banner />
+    </>
+  );
+};
+
+export default About;
